Move everything to the internal C-API for now.
