+++ Include/Python.h
@@ -84,7 +84,6 @@
 #include "iterobject.h"
 #include "cpython/initconfig.h"
 #include "pystate.h"
-#include "crossinterp.h"
 #include "cpython/genobject.h"
 #include "descrobject.h"
 #include "genericaliasobject.h"
