--- Include/crossinterp.h
@@ -1,16 +1,0 @@
-#ifndef Py_CROSSINTERP_H
-#define Py_CROSSINTERP_H
-#ifdef __cplusplus
-extern "C" {
-#endif
-
-#ifndef Py_LIMITED_API
-#  define Py_CPYTHON_CROSSINTERP_H
-#  include "cpython/crossinterp.h"
-#  undef Py_CPYTHON_CROSSINTERP_H
-#endif
-
-#ifdef __cplusplus
-}
-#endif
-#endif /* !Py_CROSSINTERP_H */
